requirements.txt
requests
beautifulsoup4
sentence-transformers
pinecone-client